pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mle-f1-tech-challenge"
version = "0.1.0"
description = "Telecommunications Customer Churn Prediction Pipeline"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "scikit-learn>=1.2.0",
    "torch>=2.0.0",
    "mlflow>=2.0.0",
    "jupyter>=1.0.0",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0"
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --verbose"
testpaths = [
    "tests",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py39"
